UBUNTU: SAUCE: (no-up) apparmor: Fix tasks not subject to, reloaded policy
This patch is against code that is not upstream.

BugLink: http://bugs.launchpad.net/bugs/1236455

When profile replacement is used the cred may not be updated for various
reasons (cached, locking ...). In these cases directly using the cred
without checking for an updated profile is wrong, and can result in
inconsistent application of old vs. new policy. This is not just an issue
of when the new policy takes over from the old but both old and new being
applied similtaneously in inconsistent ways.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>