ALSA: hda - Enable mute/mic-mute LEDs for more Thinkpads with Realtek codec
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1261296

We're using the ACPI interface to detect whether we're dealing with a Thinkpad
or not. This way we're not loading the thinkpad_acpi module when we're not on
a Thinkpad, but at the same time, we give the opportunity to check for, and
potentially enable, both present and future Thinkpad with mute/micmute LEDs.

At least those running the ALC269 family (269 to 299) of Realtek codecs.

Cc: Alex Hung <alex.hung@canonical.com>
Tested-by: Hui Wang <hui.wang@canonical.com>
Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(back ported from commit 2793769f4450d9e4845257c953ae278ecdff7d94)
Conflicts:
	sound/pci/hda/patch_realtek.c
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>