net/core: Expose bottom half of netif_receive_skb_core
Refactor __netif_receive_skb_core() in net/core/dev.c to allow
modules to invoke the handlers registered in pytpe_base.

This is useful for example in Open vSwitch when the flow action
wants to return a packet to the kernel for 'native' processing.
If the packet was punted to userspace for processing then this
action needs for the packet to be inserted into the receive
pipeline right after the rx_handler hook.

To enable this, the bottom half of __netif_receive_skb_core is
refactored into the inline __netif_receive_skb_finish() and a
new function netif_receive_skb_finish() is exported that picks
up the receive mechanism with just the skb as a parameter.